dev(shrine): more internal functions

+++ src/utils/types.cairo
@@ -3,9 +3,9 @@ use aura::utils::storage_access_impls::TroveStorageAccess;
 
 #[derive(Drop, Serde)]
 struct Trove {
-    charge_from: usize, // Time ID (timestamp // TIME_ID_INTERVAL) for start of next accumulated interest calculation
+    charge_from: u64, // Time ID (timestamp // TIME_ID_INTERVAL) for start of next accumulated interest calculation
     debt: Wad, // Normalized debt
-    last_rate_era: usize,
+    last_rate_era: u64,
 }
 
 #[derive(Drop, Serde)]
